Wire decryption into the InputStream pipeline.
This required a wholesale refactor of how streams were connected to each
other. EntryInputStream is now used to connect a decryption stream to a
decompression stream. This also (correctly) moves 'Entry level' checks
(such as CRC32) to the EntryInputStream - this makes sense as CRC32 only
exists in the context of an Entry; not the decompression stream.