Allow `OutputStream` to locally suppress Zip64 support.
Normally, while we're streaming we don't know how large a file might be
until we've seen it all, so we need to use a Zip64 extra field by default.
But if we're absolutely sure we're not going to stream more than 4GB data
into a single entry, we really don't need the Zip64 extra field - and it
can cause issues with, e.g., Microsoft office files such as xlsx and docx.

So this commit allows for the Zip64 extra field to be suppressed on an
individual `OutputStream` basis, without having to turn off Zip64
support globally.

The flag is ignored if the size of the entry is known to be larger than
4GB or if the local header offset of the entry is beyond the 4GB
address boundary within the zip file.

Closes #648